requirements.txt
amqp
argh
asgiref
async-timeout
billiard
black
boto3
botocore
Brotli
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudinary
coverage
cryptography
defusedxml
dj-database-url
Django
django-amazon-ses
django-appconf
django-autoslug
django-celery-email
django-cors-headers
django-countries
django-environ
django-filter
django-phonenumber-field
django-ses
django-storages
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djongo
djoser
dnspython
drf-yasg
exceptiongroup
factory-boy
Faker
flake8
flower
gunicorn
humanize
idna
importlib-metadata
inflection
iniconfig
isort
jmespath
kombu
Markdown
mccabe
mypy-extensions
oauthlib
packaging
pathspec
phonenumbers
pi
Pillow
platformdirs
pluggy
prometheus-client
prompt-toolkit
psycopg2-binary
pycodestyle
pycountry
pycparser
pyflakes
PyJWT
pymongo
pytest
pytest-cov
pytest-django
pytest-factoryboy
python-dateutil
python-dotenv
python3-openid
pytz
PyYAML
redis
requests
requests-oauthlib
s3transfer
six
social-auth-app-django
social-auth-core
sqlparse
stripe
tomli
tornado
typing_extensions
tzdata
uritemplate
urllib3
vine
watchdog
wcwidth
whitenoise
zipp
